refactor: Make Global Entity for ImagesEntity
