+++ dristi/lib/src/core/global_entity/global_entity.dart
@@ -1,0 +1,9 @@
+class ImagesEntity {
+  ImagesEntity({
+    required this.url,
+    required this.credit,
+  });
+
+  String url;
+  String credit;
+}
